marcketplace_microservices/products_api/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==5.1
djangorestframework==3.15.2
pillow==10.4.0
psycopg2==2.9.9
redis==5.0.8
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
